Improve sqlexec implementation
This commit makes a few fixes to the sqlexec implementation to make it
more robust.

The handler truncates the results if there are too many characters in
the message. This fixes the seemingly random "interaction failed"
errors.

The interaction uses unique IDs for each interaction so that buttons on a
previous interaction do not clobber the state of a new interaction.

This commit also limits the number of rows fetched to 500, as a
quick-and-dirty way of preventing users from blowing up our memory
usage.

Finally, this change fixes an off-by-one error in the pagination and
shuffles around some code in order to not prime the loop.